# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bluestar Alliance recruitment platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Job {
  id: number;
  title: string;
  location: string;
  category: string;
  experienceLevel: string;
  description: string;
  /** @nullable */
  salaryRange?: string | null;
  isUrgent: boolean;
  createdAt: string;
}

export interface JobInput {
  /** @minLength 2 */
  title: string;
  /** @minLength 2 */
  location: string;
  /** @minLength 2 */
  category: string;
  /** @minLength 2 */
  experienceLevel: string;
  /** @minLength 10 */
  description: string;
  salaryRange?: string;
  isUrgent?: boolean;
}

export interface ApplicationInput {
  /** @minLength 2 */
  fullName: string;
  email: string;
  phone: string;
  country: string;
  position: string;
  /** @minimum 0 */
  yearsOfExperience: number;
  coverLetter?: string;
  /** @nullable */
  cvFileName?: string | null;
}

export interface Application {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  country: string;
  position: string;
  yearsOfExperience: number;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  cvFileName?: string | null;
  status: string;
  submittedAt: string;
}

export interface ApplicationStatusUpdate {
  status: string;
}

export interface Message {
  id: number;
  applicationId: number;
  senderRole: string;
  senderName: string;
  content: string;
  isRead: boolean;
  createdAt: string;
}

export interface MessageInput {
  senderRole: string;
  senderName: string;
  /** @minLength 1 */
  content: string;
  email?: string;
}

export interface Notification {
  id: number;
  recipientEmail: string;
  recipientRole: string;
  message: string;
  type: string;
  /** @nullable */
  relatedId?: number | null;
  isRead: boolean;
  createdAt: string;
}

export interface PaymentSettings {
  id: number;
  bankName: string;
  accountName: string;
  accountNumber: string;
  /** @nullable */
  routingNumber?: string | null;
  /** @nullable */
  swiftCode?: string | null;
  /** @nullable */
  additionalInfo?: string | null;
  updatedAt: string;
}

export interface PaymentSettingsInput {
  /** @minLength 2 */
  bankName: string;
  /** @minLength 2 */
  accountName: string;
  /** @minLength 4 */
  accountNumber: string;
  routingNumber?: string;
  swiftCode?: string;
  additionalInfo?: string;
}

export interface AddonOrder {
  id: number;
  applicationId: number;
  applicantEmail: string;
  applicantName: string;
  visaSponsorship: boolean;
  flightTicket: boolean;
  workPermit: boolean;
  totalAmount: number;
  paymentMethod: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface AddonOrderInput {
  applicationId: number;
  applicantEmail: string;
  applicantName: string;
  visaSponsorship: boolean;
  flightTicket: boolean;
  workPermit: boolean;
  totalAmount: number;
  paymentMethod?: string;
  notes?: string;
}

export interface Testimonial {
  id: number;
  name: string;
  role: string;
  country: string;
  quote: string;
  avatarUrl: string;
}

export interface UserRegistration {
  email: string;
  /** @minLength 6 */
  password: string;
  fullName: string;
}

export interface UserLogin {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  fullName: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface StatsSummary {
  totalApplications: number;
  totalJobs: number;
  urgentJobs: number;
  pendingApplications: number;
  approvedApplications: number;
}

export interface RoleCount {
  position: string;
  count: number;
}

export type ListJobsParams = {
category?: string;
location?: string;
urgent?: boolean;
};

export type ListApplicationsParams = {
role?: string;
status?: string;
};

export type DeleteJob200 = {
  success: boolean;
};

export type UpdateAddonOrderStatusBody = {
  status: string;
};

export type GetMessagesParams = {
email?: string;
};

export type GetNotificationsParams = {
email?: string;
role?: string;
};

export type MarkAllNotificationsReadBody = {
  email: string;
};

export type MarkAllNotificationsRead200 = {
  success: boolean;
};

export type LogoutUser200 = {
  message: string;
};