# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bluestar Alliance recruitment platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all job openings
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: urgent
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
  /applications:
    post:
      operationId: submitApplication
      tags: [applications]
      summary: Submit a job application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "400":
          description: Validation error
  /admin/applications:
    get:
      operationId: listApplications
      tags: [admin]
      summary: Admin - List all applications
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
        "401":
          description: Unauthorized
  /admin/applications/{id}/status:
    patch:
      operationId: updateApplicationStatus
      tags: [admin]
      summary: Update application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
        "404":
          description: Not found
  /admin/jobs:
    post:
      operationId: createJob
      tags: [admin]
      summary: Admin - Create a new job listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          description: Unauthorized
  /admin/jobs/{id}:
    patch:
      operationId: updateJob
      tags: [admin]
      summary: Admin - Update a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    delete:
      operationId: deleteJob
      tags: [admin]
      summary: Admin - Delete a job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
        "404":
          description: Not found
  /admin/payment-settings:
    get:
      operationId: getPaymentSettings
      tags: [admin]
      summary: Admin - Get payment settings
      responses:
        "200":
          description: Payment settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSettings"
        "404":
          description: Not configured yet
    put:
      operationId: updatePaymentSettings
      tags: [admin]
      summary: Admin - Update payment settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentSettingsInput"
      responses:
        "200":
          description: Updated payment settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSettings"
...
  /admin/addon-orders/{id}/status:
    patch:
      operationId: updateAddonOrderStatus
      tags: [admin]
      summary: Admin - Update addon order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
      responses:
        "200":
          description: Updated addon order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddonOrder"
  /addon-orders:
    post:
      operationId: createAddonOrder
      tags: [addons]
      summary: Create an add-on order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddonOrderInput"
      responses:
        "201":
          description: Add-on order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddonOrder"
  /messages/{applicationId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for an application
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
        - name: email
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message in an application thread
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, location, category, experienceLevel, description, isUrgent, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: string
        category:
          type: string
        experienceLevel:
          type: string
        description:
          type: string
        salaryRange:
          type: ["string", "null"]
        isUrgent:
          type: boolean
        createdAt:
          type: string
    JobInput:
      type: object
      required: [title, location, category, experienceLevel, description]
      properties:
        title:
          type: string
          minLength: 2
        location:
          type: string
          minLength: 2
        category:
          type: string
          minLength: 2
        experienceLevel:
          type: string
          minLength: 2
        description:
          type: string
          minLength: 10
        salaryRange:
          type: string
        isUrgent:
          type: boolean
    ApplicationInput:
      type: object
      required: [fullName, email, phone, country, position, yearsOfExperience]
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        position:
          type: string
        yearsOfExperience:
          type: integer
          minimum: 0
        coverLetter:
          type: string
        cvFileName:
          type: ["string", "null"]
    Application:
      type: object
      required: [id, fullName, email, phone, country, position, yearsOfExperience, status, submittedAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        position:
          type: string
        yearsOfExperience:
          type: integer
        coverLetter:
          type: ["string", "null"]
        cvFileName:
          type: ["string", "null"]
        status:
          type: string
        submittedAt:
          type: string
    ApplicationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
    Message:
      type: object
      required: [id, applicationId, senderRole, senderName, content, isRead, createdAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        senderRole:
          type: string
        senderName:
          type: string
        content:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
    MessageInput:
      type: object
      required: [senderRole, senderName, content]
      properties:
        senderRole:
          type: string
        senderName:
          type: string
        content:
          type: string
          minLength: 1
        email:
          type: string
...
    PaymentSettings:
      type: object
      required: [id, bankName, accountName, accountNumber, updatedAt]
      properties:
        id:
          type: integer
        bankName:
          type: string
        accountName:
          type: string
        accountNumber:
          type: string
        routingNumber:
          type: ["string", "null"]
        swiftCode:
          type: ["string", "null"]
        additionalInfo:
          type: ["string", "null"]
        updatedAt:
          type: string
    PaymentSettingsInput:
      type: object
      required: [bankName, accountName, accountNumber]
      properties:
        bankName:
          type: string
          minLength: 2
        accountName:
          type: string
          minLength: 2
        accountNumber:
          type: string
          minLength: 4
        routingNumber:
          type: string
        swiftCode:
          type: string
        additionalInfo:
          type: string
    AddonOrder:
      type: object
      required: [id, applicationId, applicantEmail, applicantName, visaSponsorship, flightTicket, workPermit, totalAmount, paymentMethod, status, createdAt]
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        applicantEmail:
          type: string
        applicantName:
          type: string
        visaSponsorship:
          type: boolean
        flightTicket:
          type: boolean
        workPermit:
          type: boolean
        totalAmount:
          type: integer
        paymentMethod:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    AddonOrderInput:
      type: object
      required: [applicationId, applicantEmail, applicantName, visaSponsorship, flightTicket, workPermit, totalAmount]
      properties:
        applicationId:
          type: integer
        applicantEmail:
          type: string
        applicantName:
          type: string
        visaSponsorship:
          type: boolean
        flightTicket:
          type: boolean
        workPermit:
          type: boolean
        totalAmount:
          type: integer
        paymentMethod:
          type: string
        notes:
          type: string
...