# lib/api-zod/src/generated/types/addonOrderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bluestar Alliance recruitment platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AddonOrderInput {
  applicationId: number;
  applicantEmail: string;
  applicantName: string;
  visaSponsorship: boolean;
  flightTicket: boolean;
  workPermit: boolean;
  totalAmount: number;
  paymentMethod?: string;
  notes?: string;
}