# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bluestar Alliance recruitment platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  applicationId: number;
  senderRole: string;
  senderName: string;
  content: string;
  isRead: boolean;
  createdAt: string;
}