# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bluestar Alliance recruitment platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  senderRole: string;
  senderName: string;
  /** @minLength 1 */
  content: string;
  email?: string;
}