# lib/api-zod/src/generated/types/paymentSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bluestar Alliance recruitment platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentSettings {
  id: number;
  bankName: string;
  accountName: string;
  accountNumber: string;
  /** @nullable */
  routingNumber?: string | null;
  /** @nullable */
  swiftCode?: string | null;
  /** @nullable */
  additionalInfo?: string | null;
  updatedAt: string;
}