# lib/api-zod/src/generated/types/paymentSettingsInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bluestar Alliance recruitment platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentSettingsInput {
  /** @minLength 2 */
  bankName: string;
  /** @minLength 2 */
  accountName: string;
  /** @minLength 4 */
  accountNumber: string;
  routingNumber?: string;
  swiftCode?: string;
  additionalInfo?: string;
}